% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstratal.R
\name{al_nh}
\alias{al_nh}
\title{Optimal sample allocation for given strata}
\usage{
al_nh(
  gr,
  data,
  L,
  cc,
  takeall = TRUE,
  cv = FALSE,
  method = "rnabox",
  min_size = 2,
  ratios = NULL
)
}
\arguments{
\item{gr}{- strata boundaries, as vector concatenated with (L-1) values for each variable}

\item{data}{- matrix with considered variables (in columns)}

\item{L}{- number of strata}

\item{cc}{- vector of given coefficients of variation for stratification variables}

\item{takeall}{- if TRUE take-all stratum is used (default)}

\item{cv}{- if TRUE then coefficients of variations for variables for obtained allocations are printed}

\item{method}{- algorithm used for optimal allocation: 'rnabox' (default) or
'capacity'; method 'rnabox' from *stratallo* package is based on generalization of
Neyman optimal allocation formula taking into account lower an upper bounds for
sample sizes in strata; method 'capacity' uses integer optimal allocation
'CapacityScaling' algorithm from Friedrich et al. (2015) paper}

\item{min_size}{- minimal sample size in strata (default 2); if min_size < 1 then minimal
sample fraction in strata}

\item{ratios}{- vector estimates for ratios, used when stratified variables
are given after linearization}
}
\value{
numerical vector with optimal sample allocation in strata
defined by input boundaries;
}
\description{
Function for optimal sample allocation in L strata defined by boundaries
}
\examples{

# Generation of correlated lognormal variables (x,y)
set.seed(3456)
ro <- 0.5 # correlation coefficient for lognormal variables
#
# correlation coefficient for normal variables (theoretical formula)
(ro_norm <- log(.5 * (exp(1) - 1) + 1))
#
x <- rnorm(10000)
z <- rnorm(10000)
y <- x * ro_norm + z * sqrt(1 - ro_norm^2)
x <- exp(x)
y <- exp(y)
cor(x, y)
#
dataxy <- cbind(x, y) # data matrix
L <- 3 # number of strata
cv <- 0.05 # coefficient of variation for estimators
boundaries <- c(cumfp(dataxy[, 1], L, 0.5), cumfp(dataxy[, 2], L, 0.5))
nh <- al_nh(boundaries, # strata boundaries
  data = dataxy,
  L = L,
  cc = c(cv, cv),
  cv = TRUE,
  method = "rnabox"
)
cat("Optimal sample allocation for given strata = ", nh, "\n")
cat("Optimal sample size = ", sum(nh), "\n")
#
nh <- al_nh(boundaries,
  data = dataxy,
  L = L,
  cc = c(cv, cv),
  cv = TRUE,
  method = "rnabox",
  min_size = 20 # change for minimal sample size
)
cat("Optimal sample allocation for given strata = ", nh, "\n")
cat("Optimal sample size = ", sum(nh), "\n")

}
\references{
- Ulf Friedrich, Ralf Münnich, Sven de Vries, Matthias Wagner,
Fast integer-valued algorithms for optimal allocations under constraints in stratified sampling,
Computational Statistics and Data Analysis 92 (2015), 1-12

- Wesolowski J., Wieczorkowski R., Wojciak W. (2024),
Recursive Neyman Algorithm for Optimum Sample Allocation under Box Constraints on
Sample Sizes in Strata, (to be published in *Survey Methodology*),
arxiv version available at: \url{https://arxiv.org/abs/2304.07034}
}
