% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstratal.R
\name{cumfp}
\alias{cumfp}
\title{Cumulative power density rule for initial stratification}
\usage{
cumfp(x, L, p)
}
\arguments{
\item{x}{- stratification numerical variable}

\item{L}{- number of strata}

\item{p}{- value of power,  from interval (0,1)}
}
\value{
initial strata boundaries
}
\description{
Function for computation initial stratification by generalized
cumroot density rule i.e. cumulative power density rule
}
\examples{
# Examples for cumfp
# generation of artificial data
x1 <- rchisq(10000, 1)

cumfp(x = x1, L = 5, p = 0.5) # classical cumroot rule

cumfp(x = x1, L = 5, p = 0.3)

}
